+++ drinking_buddies/drinking_auth/models.py
@@ -78,7 +78,7 @@ class User(AbstractBaseUser, PermissionsMixin):
     is_staff = models.BooleanField(default=False)
     is_superuser = models.BooleanField(default=False)
     is_active = models.BooleanField(default=True)
-    last_login = models.DateTimeField(null=False, blank=False)
+    last_login = models.DateTimeField(null=True, blank=False)
     date_joined = models.DateTimeField(auto_now_add=True)
 
     USERNAME_FIELD = "username"