+++ drinking_buddies/schema.py
@@ -1,6 +1,7 @@
 """Schemas for all project applications."""
 import graphene
 
+from drinking_buddies.mutations import Mutation
 from drinking_buddies.query import Query
 
-schema = graphene.Schema(query=Query)
+schema = graphene.Schema(query=Query, mutation=Mutation)